+++ src/main/kotlin/se/uulm/snowballr/backend/model/exception/notfound/StageNotFoundException.kt
@@ -6,4 +6,4 @@ import se.uulm.snowballr.backend.model.exception.NotFoundException
 /**
  * Represents an exception that occurs when a stage in a [Project] is not found.
  */
-class StageNotFoundException(stage: Long) : NotFoundException("Stage '$stage' not found.")
+class StageNotFoundException(stage: Int) : NotFoundException("Stage '$stage' not found.")
